requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
pandas>=1.5.0
python-dotenv>=0.21.0
gspread>=5.7.0
google-auth>=2.15.0
